+++ sdk/eventgrid/azure-messaging-eventgrid-cloudnative-cloudevents/CHANGELOG.md
@@ -3,4 +3,5 @@
 ## 1.0.0-beta.1 (Unreleased)
 
 ### Features Added
-- Bridge library to allow sending CloudNative CloudEvents using Azure Event Grid.
+- Added the bridge library to allow sending Cloud Native Computing Foundation(CNCF) CloudEvents using Azure Event Grid.
+  More information about CNCF cloudEvents see https://cloudevents.github.io/sdk-java/.